// src/app/models/movie.model.ts
export class Movie {
  backdrop_path: string = "";
  first_air_date: string = "";
  release_date: string = "";
  id: number = 0;
  media_type: string = "";
  original_name: string = "";
  original_title: string = "";
  overview: string = "";
  poster_path: string = "";
  isMovie: boolean = false;
  hasTrailer: boolean = false;
  vote_average: number = 0;
}